Rename repo browse file to 'search'
